requirements.txt
Django==4.0.6
PyGithub
pathspec-0.9.0
